requirements.txt
pytorch == 1.8.0
spatial-correlation-sampler >=0.2.1